Add files via upload
This is the code of the first exercise of chapter 7.
Its purpose is to verify if a password is a "strong password".
It is to verify 3 criteria:
1. The length of the password is not less than 8 characters.
2. The password needs to have both uppercase and lowercase letters.
3. The password needs to have at least 1 number.
It has to be done with regular expressions.

My code has some issues. If you could teach me how to improve it, I will be very appreciate.